DOPID assembly function
These routines calculate the new PWM duty cycle using the MC68HC11N4 math coprocessor.  The code can be run on an M68HC11EVS with an M68HC11K4 emulator and MC68HC11N4 processor.  The EVS monitor should be 1.1 or later. The EVS and vectors were set to SPECIAL TEST MODE to aid debug.  This code is called by a C routine but could be converted to an all assembly environment by defining the variables in assembly instead of as externals.